Changed some variable declarations

+++ src/lib/examples/PID.h
@@ -3,14 +3,15 @@
 
 class PID{
   private:
-    int Kp, Ki, Kd;
+    double Kp, Ki, Kd;
     double P, I, D, E;
 
   public:
     PID();
-    PID(int, int, int);
+    PID(double, double, double);
     int pid_int(int*);
     int pid_bool(int*);
 };
 
 #endif
+
